refactor(services): move balance_query_service to api/
